Added new changes and resolve all the comments succesfully. Waiting for merge pull request
