+++ App.tsx
@@ -1,10 +1,6 @@
 import React from 'react';
+import Criptotracker from './components/Criptotracker';
 
-
-import Criptotracker from './components/Criptotracker/Criptotracker';
-
-const App(): JSX.Element = () => (
-    <Criptotracker />
-);
+const App: Element = () => <Criptotracker />;
 
 export default App;
