+++ components/Criptotracker/styles.ts
@@ -1,11 +1,7 @@
-/* eslint-disable prettier/prettier */
 import styled from 'styled-components/native';
-
-import { theme } from '../../utils/theme';
+import {theme} from '../../utils/theme';
 
 export const PrincipalView = styled.SafeAreaView`
   background-color: ${theme.colors.white};
   flex: 1;
 `;
-
-
